First Commit
This is the first commit after initializing the `osdev` repo. The repo
was created after quite a bit of progress had already been made on the
kernel. This commit includes the `boot` directory -- the directory of my
custom bootloader -- for historical reasons only. I migrated to GRUB a
while ago for many reasons, therefore the `boot` directory may be
removed from future commits.

Notes:

- The kernel is a "higher-half" kernel
- The current drivers included are:
  - A screen driver
  - A serial driver
  - A keyboard driver
  - An ata driver
- No filesystem yet (ext2 driver planned)
- The memory manager is in progress (a custom buddy-allocator)
- There is primitive "pre-allocator" watermark allocator
- There is a fairly competent `printf` implementation
- A **lot** of work still needs to be done -- see `todo.txt`