Switched project to Makefiles
Among a long list of other changes, the project now uses Makefiles
instead of CMake for the build process. Makefiles were needed for a
large part of the build process before, so it made the most sense
just to fully commit to using Makefiles for the entire project.
Also, the memory allocator is now finished. It uses recursive page
mapping, and now works for allocating and freeing pages of memory.
Work is now being done on the ext2 filesystem driver along with the
virtual filesystem.