Scheduler v2 - Added thread support
Before now, the scheduler operated on processes which meant you had to
call `process_fork()` (a relatively expensive call) when writing things
like drivers. As a solution to this problem, this commit adds proper
kernel level threads with a pthreads-like api, in addition to mutex and
condition variable data structures. The scheduler now schedules threads
exlusively, as each process allocates a "main" thread on creation. In
addition to everything else, `lib/lock.c` was removed in favor of a
spinlock implementation that lives in the `kernel/` folder. This marks
a huge milestone as threads can now be used in the many kernel level
drivers to come.