fixed xhci control transfers
this commit fixes a ton of small bugs with the xhci driver code.
these were mainly caused by small mistakes in the data structure
or trb definitions. also, control transfers now work both in part
to some of the fixes mentioned above as well as the addition of a
status stage which is required by the spec.