Restructured xhci driver code
This commit restructures the xhci code so that it is now all contained
in a single xhci.c file. The lines between software and hardware are a
little mixed up so it makes more sense for all the code to live in one
place. Since there are so many spec related definitions, those still
live in a seperate xhci_hw.h header file.