Mutex/condition rewrite
This commit is a total rewrite of the mutex and condition locking
primitives. It reduces the total size of both structures, removes the
use of dynamic memory allocation and generally improves performance.
Before, each thread that was blocked by a mutex or condition would
have a thread_link_t struct allocated so it could be tracked without
messing up the next and prev pointers used by the scheduler blocked
list. This commit adds a thread flag that allows other parts of the code
to maintain ownership of the thread when blocked and prevents the
scheduler from touching those pointers during that time. This allows us
to get rid of the thread_link_t struct and use the built-in next and
prev pointers on the thread itself in order to track the blocked
threads.