Filesystem v2
This is a big commit, mainly because im bad at making procedural commits
as I go along. This is a complete rewrite of the entire filesystem and
virtual filesystem. The original one was written before other things
were added, and it quickly became clear that it wasn't flexible enough
to properly handle complex devices and on-disk filesystems. A lot of
inspiration for how stuff is layed out is taken from Linux, although
this is much, much simpler. This is just the first step, following
commits will be made to implement the FAT filesystem driver using this
new api, as well as adding special device files like /dev/null and
/dev/zero (among other things).