ELF loading + dynamic linking v1
This commit adds support for loading elf files compiled with the os
toolchain. Programs compiled by the toolchain require the runtime linker
ld.so to itself be dynamically linked with the program. This commit
includes changes to loader.c which facilitate the loading and parsing of
the elf program and the linker. Addtionally, process_execve has been
rewritten to construct the proper entry stack for the linker according
to x86_64 System V ABI. Further work must be done to implement the
system calls required by the linker but you can successfully sysret to
the linker code.