Added custom meta build system for userland targets
This commit adds a custom meta build system powered by m4 macros. This
was needed because defining new targets residing in the sys/ directory
was becoming more difficult, especially when the targets could be
exectuables or static/dynamic libraries, each having their own way of
linking and also installing in the system root. The scripts/macros.m4
file defines all of the macros used for defining targets, and right now
the Makefile must be generated from the Makefile.m4 manually.