Reverted individual process space change
It dawned on me that having processes spaces seperate like that means
that the shared high-half would go out of sync between process vm trees.
For now this change has been reverted, but in the future we need to find
a way to have parts of an interval tree shared, or have two trees: one
for the low half (per process) and a single shared tree for the upper
half of memory.