re-disabled smp initialization and scheduling
Even though multi-processor scheduling does somewhat work, there are
many race conditions in various parts of the kernel that resulted in
freezes or panics and general instability. I will re-disable it for now
but plan to add better logging and tools so that debugging things with
concurrency enabled might be a bit easier.