fs rewrite part 1 - specification
While working on the initrd support I became annoyed with a number of
things in the old fs code. The main problem was that the internal API
for implementing a new filesystem was inconsistent, poorly documented
and overly complex. I'm setting out to rewrite the internal and external
filesystem APIs with to better serve the os in the long term.

This commit contains the initial fs_types.h file which defines and
documents the various types and operations used in the filesystem.
It has detailed doc comments describing the operations that need to be
implemented by filesystems, what they should do and what they can and
can not modify. This will certainly change, but having the API design
and layout mostly documented is good start.