fs rewrite part 3 - devices
This is a big commit that includes way too much, sorry. The biggest
thing added is the new kernel device core. I realized that I needed
to fully implement devices in order to continue with the filesystem and
it was also sorely needed. The new device system manages buses, drivers
and devices on the system. I also wrote pci drivers in order to
test-drive the interfaces and api in the wild. I will be converting some
others to the new driver system, but stuff like USB will take longer.