fs rewrite v3 - complete redesign for multi-cpu systems
I realised a little late into rewrite v2 that it fundamentally wouldnt
work in a multi-cpu context and so I deleted it and started from the
beginning. This time I read a lot about the Solaris and FreeBSD vfs
implementations and followed the same basic structure, although I did
diverge in a number of ways. All the major structures are reference
counted and there is locking everywhere. This certainly is riddled with
bugs but this is just the beginning.